+++ Calculator/Core/Inc/stack.h
@@ -15,16 +15,16 @@
 struct stack {
     uint32_t maxsize;
     int32_t top;
-    float *data;
+    double *data;
 };
 
 struct stack *init_stack(uint32_t size);
 uint32_t stack_size(struct stack *st);
 int8_t stack_isempty(struct stack *st);
 int8_t stack_isfull(struct stack *st);
-int8_t push(struct stack *st, float val);
-float pop(struct stack *st);
-float peek(struct stack *st);
+int8_t push(struct stack *st, double val);
+double pop(struct stack *st);
+double peek(struct stack *st);
 void free_stack(struct stack *st);
 
 #endif //STACK_H
